requirements.txt
celery==5.3.0
importlib-metadata==6.6.0 
Django>=4.2.1
psycopg2-binary>=2.9.6
gunicorn>=20.1.0
redis==4.5.5
python-decouple==3.8
cached-property==1.5.2 
flower==1.2.0